refactor: move more stuff in own src
